new option assignment syntax
